apps/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
google-genai>=1.0.0
PyJWT>=2.8.0
